feat: First bit of DB stuff!

+++ api/src/infrastructure/paging.rs
@@ -2,5 +2,5 @@ use serde::Serialize;
 
 #[derive(Serialize)]
 pub struct Paging {
-    pub total: u32,
+    pub total: usize,
 }
